refactor: SelectItem 재사용을 위한 파일 분리
